backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
httpx
appwrite
apscheduler
python-dotenv